Rename String due to PHP 7 conflict
